requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pinecone==5.0.0
google-genai==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
pydantic>=2.10,<3
dotenv

tenacity